Remove Google.Play.Billing.asmdef file
Version 2.1.0 of the "com.unity.purchasing" package switched from DLL to source distribution, which changed how Google.Play.Billing.asmdef needs to be configured. Since the asmdef is required when building a .tgz file for Unity Package Manager, but not required for .unitypackage files, remove it from the project and only include it when building a .tgz file.

PiperOrigin-RevId: 333141303